app/requirements.txt
api_analytics==1.2.3
extra_streamlit_components==0.1.71
fastapi==0.111.1
passlib==1.7.4
pika==1.3.2
pydantic
pydantic_settings==2.4.0
pytest==8.3.2
python-dotenv==1.0.1
python_jose==3.3.0
Requests==2.32.3
sqlmodel==0.0.21
streamlit==1.36.0
uvicorn==0.17.4
psycopg
joblib
scikit-learn
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
jiter==0.7.1
openai==1.54.4
google-generativeai